LIBITD-281. Moved XSL to Filesystem.
Added an XSLGetter class which reads the XSL file from ../xsls/* from the web app base. This way recompilation of web app won’t be needed just to change the xsl file. A sample XSL is included in the source code (at the root - sample.xsl) for developer’s reference.

https://issues.umd.edu/browse/LIBITD-281