LIBITD-294. Adding new parameter for incorrect fixing download URL.
The download URL was set to localhost:8080 in the code, which was not allowing us to download any XSLs from the server, so I added a new parameter "xslDownloadURL" in the web.xml, so that it can be changed while deploying to new server, without recompiling the code. We can set the new URL in server's context.xml and set that one to override this web.xml configuration.
Instructions will be updated in the documentation. Currently it should be set to "https://wwwdev.lib.umd.edu/wufoo-connector/xsls/" in the context.xml.

https://issues.umd.edu/browse/LIBITD-294